Revert "Removed check on request and removed stars on routes of product"
This reverts commit 17d86ce0a69e3fca8f7418d256912cf544d93018.

+++ backend/controllers/faq_controller.js
@@ -19,6 +19,8 @@ class Faq_controller{
 
     async newFaq(bodyFE){
 
+        if (!bodyFE.newQuestion || !bodyFE.newAnswer) return [500, "Errore, richiesta non formulata correttamente!"];
+
         const newFaq = await Database.faq.create({
             
             question: bodyFE.newQuestion,
@@ -31,6 +33,8 @@ class Faq_controller{
     }
     
     async updateFaq(bodyFE){
+
+        if (!bodyFE.question || !bodyFE.answer) return [500, "Errore, richiesta non formulata correttamente!"];
 
         const editedFaq = await Database.faq.update(
             { 
@@ -50,6 +54,8 @@ class Faq_controller{
     }
     
     async deleteFaq(bodyFE){
+
+        if (!bodyFE.id) return [500, "Errore, richiesta non formulata correttamente!"];
 
         const deletedFaq = await Database.faq.destroy({
             where: {
